API to get the booking list given the user ID. (#9)
getListaPrenotazioniUtente(idUtente) returns the list of booking associated with it, in particular the attributes:
'data'
'oraInizio'
'oraFine'
'nome'
'città'
'indirizzo'
N.B.: I would insert it since when we implement the cancellation later the user must be able to see all the reservations he made in any field. For the moment it is a draft because we should check that the reservation is not a past one.